requirements.txt
Django==1.9
django-admin-tools==0.8.1
django-datatable==0.3.1
django-tastypie==0.14.2
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
simpy==3.0.11
six==1.11.0
